fix: handle the branch-before-tag case
Consider this tree. We want to know what to tag HEAD, which is at the
tip of [main]:

  M─┐ [main] chore: merge again
  │ o [new-branch-1] fix: another minor change
  M─│─┐ <v0.1.1> chore: merge
  │ │ o [new-branch-2] fix: minor change
  I─┴─┘ <v0.1.0> chore: initial commit

The [main] parent of the current HEAD contains these commits:

  * chore: merge again
  * <v0.1.1> chore: merge

The [new-branch-1] parent of the current HEAD contains these commits:

  * chore: merge again
  * fix: another minor change
  * <v0.1.0> chore: initial commit

Either history considered by itself indicates the new tag for HEAD
should be v0.1.1. But that is not correct, because the histories need to
be _combined_ to determine the correct version: v0.1.2.

This is the scenario that this commit fixes.